Moved various store methods to separate modules for readability and easier upgrades
