+++ .gitignore
@@ -130,3 +130,5 @@ dmypy.json
 .pyre/
 
 .vscode/
+
+*.json
