+++ requirements.txt
@@ -1,3 +1,4 @@
 discord.py
 discord-py-slash-command
 requests
+python-dotenv
